Moved authentication to CLI part
This moves all of the OneNote authentication into the CLI by starting up
an express server, and listening for the redirect to localhost.  It's
probably a bad solution, but I don't know of a better one.

I use the user's default browser to open up the OneNote authentication
page, and then they can authorize it.  Once it's authorize, I snag the
token and shut the server down after redirecting the user to a landing
page.

Things to do for authentication still:
- Save the auth data to a file.
- Make a real landing page for the user.
- Handle expiry of auth code (though, honestly, I probably won't do this
  for a while.)
- Write better promises, probably?